BE-Feat: express 서버, db 연결 컴포넌트 추가
